fix: rename typography examples [MDS-328]
