+++ next-docs/public/examples/typography/Caption.tsx
